+++ pkg/pxc/queries/queries.go
@@ -1,17 +1,20 @@
 package queries
 
 import (
+	"bytes"
 	"context"
 	"database/sql"
+	"encoding/csv"
 	"fmt"
+	"regexp"
 	"strings"
 
 	"github.com/pkg/errors"
 
-	"github.com/go-sql-driver/mysql"
+	"github.com/gocarina/gocsv"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/types"
-	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	"github.com/percona/percona-xtradb-cluster-operator/clientcmd"
 )
 
 type ReplicationStatus int8
@@ -31,10 +34,6 @@ const (
 // https://github.com/percona/percona-docker/blob/pxc-operator-1.3.0/proxysql/dockerdir/etc/proxysql-admin.cnf#L23
 const writerID = 11
 
-type Database struct {
-	db *sql.DB
-}
-
 type ReplicationConfig struct {
 	Source             ReplicationChannelSource
 	SourceRetryCount   uint
@@ -53,51 +52,77 @@ type ReplicationChannelSource struct {
 
 var ErrNotFound = errors.New("not found")
 
-func New(client client.Client, namespace, secretName, user, host string, port int32, timeout int32) (Database, error) {
-	secretObj := corev1.Secret{}
-	err := client.Get(context.TODO(),
-		types.NamespacedName{
-			Namespace: namespace,
-			Name:      secretName,
-		},
-		&secretObj,
-	)
-	if err != nil {
-		return Database{}, err
-	}
+var sensitiveRegexp = regexp.MustCompile(":.*@")
 
-	timeoutStr := fmt.Sprintf("%ds", timeout)
-	config := mysql.NewConfig()
-	config.User = user
-	config.Passwd = string(secretObj.Data[user])
-	config.Net = "tcp"
-	config.DBName = "mysql"
-	config.Addr = fmt.Sprintf("%s:%d", host, port)
-	config.Params = map[string]string{
-		"interpolateParams": "true",
-		"timeout":           timeoutStr,
-		"readTimeout":       timeoutStr,
-		"writeTimeout":      timeoutStr,
-		"tls":               "preferred",
-	}
-
-	db, err := sql.Open("mysql", config.FormatDSN())
-	if err != nil {
-		return Database{}, err
-	}
-
-	err = db.Ping()
-	if err != nil {
-		return Database{}, err
-	}
-
-	return Database{
-		db: db,
-	}, nil
+type Database struct {
+	client    *clientcmd.Client
+	pod       *corev1.Pod
+	container string
+	cmd       []string
 }
 
-func (p *Database) CurrentReplicationChannels() ([]string, error) {
-	rows, err := p.db.Query(`SELECT DISTINCT(Channel_name) from replication_asynchronous_connection_failover`)
+// NewPXC creates a new Database instance for a given PXC pod
+func NewPXC(pod *corev1.Pod, cliCmd *clientcmd.Client, user, pass, host string) *Database {
+	cmd := []string{"mysql", "--database", "mysql", fmt.Sprintf("-p%s", pass), "-u", string(user), "-h", host, "-e"}
+	return &Database{client: cliCmd, pod: pod, container: "pxc", cmd: cmd}
+}
+
+// NewProxySQL creates a new Database instance for a given ProxySQL pod
+func NewProxySQL(pod *corev1.Pod, cliCmd *clientcmd.Client, user, pass string) *Database {
+	cmd := []string{"mysql", fmt.Sprintf("-p%s", pass), "-u", string(user), "-h", "127.0.0.1", "-P", "6032", "-e"}
+	return &Database{client: cliCmd, pod: pod, container: "proxysql", cmd: cmd}
+}
+
+// NewHAProxy creates a new Database instance for a given HAProxy pod
+func NewHAProxy(pod *corev1.Pod, cliCmd *clientcmd.Client, user, pass string) *Database {
+	cmd := []string{"mysql", fmt.Sprintf("-p%s", pass), "-u", string(user), "-h", "127.0.0.1", "-e"}
+	return &Database{client: cliCmd, pod: pod, container: "haproxy", cmd: cmd}
+}
+
+func (d *Database) exec(ctx context.Context, stm string, stdout, stderr *bytes.Buffer) error {
+	cmd := append(d.cmd, stm)
+	err := d.client.Exec(d.pod, d.container, cmd, nil, stdout, stderr, false)
+	if err != nil {
+		sout := sensitiveRegexp.ReplaceAllString(stdout.String(), ":*****@")
+		serr := sensitiveRegexp.ReplaceAllString(stderr.String(), ":*****@")
+		return errors.Wrapf(err, "run %s, stdout: %s, stderr: %s", cmd, sout, serr)
+	}
+
+	if strings.Contains(stderr.String(), "ERROR") {
+		return fmt.Errorf("sql error: %s", stderr)
+	}
+
+	return nil
+}
+
+func (d *Database) query(ctx context.Context, query string, out interface{}) error {
+	var errb, outb bytes.Buffer
+	err := d.exec(ctx, query, &outb, &errb)
+	if err != nil {
+		return err
+	}
+
+	if !strings.Contains(errb.String(), "ERROR") && outb.Len() == 0 {
+		return sql.ErrNoRows
+	}
+
+	csv := csv.NewReader(bytes.NewReader(outb.Bytes()))
+	csv.Comma = '\t'
+
+	if err = gocsv.UnmarshalCSV(csv, out); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (d *Database) CurrentReplicationChannels(ctx context.Context) ([]string, error) {
+	rows := []*struct {
+		Name string `csv:"name"`
+	}{}
+
+	q := `SELECT DISTINCT(Channel_name) as name from replication_asynchronous_connection_failover`
+	err := d.query(ctx, q, &rows)
 	if err != nil {
 		if errors.Is(err, sql.ErrNoRows) {
 			return nil, nil
@@ -105,151 +130,133 @@ func (p *Database) CurrentReplicationChannels() ([]string, error) {
 		return nil, errors.Wrap(err, "select current replication channels")
 	}
 
-	defer rows.Close()
-
 	result := make([]string, 0)
-	for rows.Next() {
-		src := ""
-		err = rows.Scan(&src)
-		if err != nil {
-			return nil, errors.Wrap(err, "scan channel name")
-		}
-		result = append(result, src)
+	for _, row := range rows {
+		result = append(result, row.Name)
 	}
 	return result, nil
 }
 
-func (p *Database) ChangeChannelPassword(channel, password string) error {
-	tx, err := p.db.Begin()
+func (d *Database) ChangeChannelPassword(ctx context.Context, channel, password string) error {
+	var errb, outb bytes.Buffer
+
+	q := fmt.Sprintf(`START TRANSACTION;
+		STOP REPLICA IO_THREAD FOR CHANNEL %s;
+		CHANGE REPLICATION SOURCE TO SOURCE_PASSWORD='%s' FOR CHANNEL %s;
+		START REPLICA IO_THREAD FOR CHANNEL %s;
+		COMMIT;
+	`, channel, password, channel, channel)
+
+	err := d.exec(ctx, q, &outb, &errb)
 	if err != nil {
-		return errors.Wrap(err, "start transaction for updating channel password")
+		return errors.Wrap(err, "change channel password")
 	}
-	_, err = tx.Exec(`STOP REPLICA IO_THREAD FOR CHANNEL ?`, channel)
-	if err != nil {
-		errT := tx.Rollback()
-		if errT != nil {
-			return errors.Wrapf(err, "rollback STOP REPLICA IO_THREAD FOR CHANNEL %s", channel)
-		}
-		return errors.Wrapf(err, "stop replication IO thread for channel %s", channel)
-	}
-	_, err = tx.Exec(`CHANGE REPLICATION SOURCE TO SOURCE_PASSWORD=? FOR CHANNEL ?`, password, channel)
-	if err != nil {
-		errT := tx.Rollback()
-		if errT != nil {
-			return errors.Wrapf(err, "rollback CHANGE SOURCE_PASSWORD FOR CHANNEL %s", channel)
-		}
-		return errors.Wrapf(err, "change master password for channel %s", channel)
-	}
-	_, err = tx.Exec(`START REPLICA IO_THREAD FOR CHANNEL ?`, channel)
-	if err != nil {
-		errT := tx.Rollback()
-		if errT != nil {
-			return errors.Wrapf(err, "rollback START REPLICA IO_THREAD FOR CHANNEL %s", channel)
-		}
-		return errors.Wrapf(err, "start io thread for channel %s", channel)
-	}
-	return tx.Commit()
+
+	return nil
 }
 
-func (p *Database) ReplicationStatus(channel string) (ReplicationStatus, error) {
-	rows, err := p.db.Query(`SHOW REPLICA STATUS FOR CHANNEL ?`, channel)
+func (d *Database) ReplicationStatus(ctx context.Context, channel string) (ReplicationStatus, error) {
+	rows := []*struct {
+		IORunning  string `csv:"Replica_IO_Running"`
+		SQLRunning string `csv:"Replica_SQL_Running"`
+		LastErrNo  int    `csv:"Last_Errno"`
+	}{}
+
+	q := fmt.Sprintf("SHOW REPLICA STATUS FOR CHANNEL %s", channel)
+	err := d.query(ctx, q, &rows)
 	if err != nil {
-		if strings.HasSuffix(err.Error(), "does not exist.") || errors.Is(err, sql.ErrNoRows) {
+		if errors.Is(err, sql.ErrNoRows) {
 			return ReplicationStatusNotInitiated, nil
 		}
-		return ReplicationStatusError, errors.Wrap(err, "get current replica status")
+		return ReplicationStatusError, errors.Wrap(err, "select replication status")
 	}
 
-	defer rows.Close()
-	cols, err := rows.Columns()
-	if err != nil {
-		return ReplicationStatusError, errors.Wrap(err, "get columns")
-	}
-	vals := make([]interface{}, len(cols))
-	for i := range cols {
-		vals[i] = new(sql.RawBytes)
-	}
+	ioRunning := rows[0].IORunning == "Yes"
+	sqlRunning := rows[0].SQLRunning == "Yes"
+	lastErrNo := rows[0].LastErrNo
 
-	for rows.Next() {
-		err = rows.Scan(vals...)
-		if err != nil {
-			return ReplicationStatusError, errors.Wrap(err, "scan replication status")
-		}
-	}
-
-	IORunning := string(*vals[10].(*sql.RawBytes))
-	SQLRunning := string(*vals[11].(*sql.RawBytes))
-	LastErrNo := string(*vals[18].(*sql.RawBytes))
-	if IORunning == "Yes" && SQLRunning == "Yes" {
+	if ioRunning && sqlRunning {
 		return ReplicationStatusActive, nil
 	}
 
-	if IORunning == "No" && SQLRunning == "No" && LastErrNo == "0" {
+	if !ioRunning && !sqlRunning && lastErrNo == 0 {
 		return ReplicationStatusNotInitiated, nil
 	}
 
 	return ReplicationStatusError, nil
 }
 
-func (p *Database) StopAllReplication() error {
-	_, err := p.db.Exec("STOP REPLICA")
-	return errors.Wrap(err, "failed to stop replication")
+func (d *Database) StopAllReplication(ctx context.Context) error {
+	var errb, outb bytes.Buffer
+	if err := d.exec(ctx, "STOP REPLICA", &outb, &errb); err != nil {
+		return errors.Wrap(err, "failed to stop replication")
+	}
+	return nil
 }
 
-func (p *Database) AddReplicationSource(name, host string, port, weight int) error {
-	_, err := p.db.Exec("SELECT asynchronous_connection_failover_add_source(?, ?, ?, null, ?)", name, host, port, weight)
-	return errors.Wrap(err, "add replication source")
+func (d *Database) AddReplicationSource(ctx context.Context, name, host string, port, weight int) error {
+	var errb, outb bytes.Buffer
+	q := fmt.Sprintf("SELECT asynchronous_connection_failover_add_source('%s', '%s', %d, null, %d)", name, host, port, weight)
+	err := d.exec(ctx, q, &outb, &errb)
+	if err != nil {
+		return errors.Wrap(err, "add replication source")
+	}
+	return nil
 }
 
-func (p *Database) ReplicationChannelSources(channelName string) ([]ReplicationChannelSource, error) {
-	rows, err := p.db.Query(`
-        SELECT host,
-               port,
-               weight
-        FROM   replication_asynchronous_connection_failover
-        WHERE  channel_name = ?
-    `, channelName)
+func (d *Database) ReplicationChannelSources(ctx context.Context, channelName string) ([]ReplicationChannelSource, error) {
+	rows := []*struct {
+		Host   string `csv:"host"`
+		Port   int    `csv:"port"`
+		Wieght int    `csv:"weight"`
+	}{}
+
+	q := fmt.Sprintf("SELECT host, port, weight FROM replication_asynchronous_connection_failover WHERE Channel_name = '%s'", channelName)
+	err := d.query(ctx, q, &rows)
 	if err != nil {
 		if errors.Is(err, sql.ErrNoRows) {
 			return nil, ErrNotFound
 		}
 		return nil, errors.Wrap(err, "get replication channels")
 	}
-	defer rows.Close()
+
 	result := make([]ReplicationChannelSource, 0)
-	for rows.Next() {
-		r := ReplicationChannelSource{}
-		err = rows.Scan(&r.Host, &r.Port, &r.Weight)
-		if err != nil {
-			return nil, errors.Wrap(err, "read replication channel info")
-		}
-		result = append(result, r)
+	for _, row := range rows {
+		result = append(result, ReplicationChannelSource{Host: row.Host, Port: row.Port, Weight: row.Wieght})
 	}
 	return result, nil
 }
 
-func (p *Database) StopReplication(name string) error {
-	_, err := p.db.Exec("STOP REPLICA FOR CHANNEL ?", name)
+func (d *Database) StopReplication(ctx context.Context, name string) error {
+	var errb, outb bytes.Buffer
+	err := d.exec(ctx, fmt.Sprintf("STOP REPLICA FOR CHANNEL '%s'", name), &outb, &errb)
 	return errors.Wrap(err, "stop replication for channel "+name)
 }
 
-func (p *Database) EnableReadonly() error {
-	_, err := p.db.Exec("SET GLOBAL READ_ONLY=1")
+func (d *Database) EnableReadonly(ctx context.Context) error {
+	var errb, outb bytes.Buffer
+	err := d.exec(ctx, "SET GLOBAL READ_ONLY=1", &outb, &errb)
 	return errors.Wrap(err, "set global read_only param to 1")
 }
 
-func (p *Database) DisableReadonly() error {
-	_, err := p.db.Exec("SET GLOBAL READ_ONLY=0")
+func (d *Database) DisableReadonly(ctx context.Context) error {
+	var errb, outb bytes.Buffer
+	err := d.exec(ctx, "SET GLOBAL READ_ONLY=0", &outb, &errb)
 	return errors.Wrap(err, "set global read_only param to 0")
 }
 
-func (p *Database) IsReadonly() (bool, error) {
-	readonly := 0
-	err := p.db.QueryRow("select @@read_only").Scan(&readonly)
-	return readonly == 1, errors.Wrap(err, "select global read_only param")
+func (p *Database) IsReadonlyExec(ctx context.Context) (bool, error) {
+	rows := []*struct {
+		ReadOnly int `csv:"readOnly"`
+	}{}
+	err := p.query(ctx, "select @@read_only as readOnly", &rows)
+	if err != nil {
+		return false, errors.Wrap(err, "select global read_only param")
+	}
+	return rows[0].ReadOnly == 1, nil
 }
 
-func (p *Database) StartReplication(replicaPass string, config ReplicationConfig) error {
+func (d *Database) StartReplication(ctx context.Context, replicaPass string, config ReplicationConfig) error {
 	var ca string
 	var ssl int
 	if config.SSL {
@@ -262,37 +269,51 @@ func (p *Database) StartReplication(replicaPass string, config ReplicationConfig
 		sslVerify = 1
 	}
 
-	_, err := p.db.Exec(`
-	CHANGE REPLICATION SOURCE TO
-		SOURCE_USER='replication',
-		SOURCE_PASSWORD=?,
-		SOURCE_HOST=?,
-		SOURCE_PORT=?,
-		SOURCE_CONNECTION_AUTO_FAILOVER=1,
-		SOURCE_AUTO_POSITION=1,
-		SOURCE_RETRY_COUNT=?,
-		SOURCE_CONNECT_RETRY=?,
-		SOURCE_SSL=?,
-		SOURCE_SSL_CA=?,
-		SOURCE_SSL_VERIFY_SERVER_CERT=?
-		FOR CHANNEL ?
-`, replicaPass, config.Source.Host, config.Source.Port, config.SourceRetryCount, config.SourceConnectRetry, ssl, ca, sslVerify, config.Source.Name)
+	q := fmt.Sprintf(`
+		CHANGE REPLICATION SOURCE TO
+			SOURCE_USER='replication',
+			SOURCE_PASSWORD='%s',
+			SOURCE_HOST='%s',
+			SOURCE_PORT=%d,
+			SOURCE_CONNECTION_AUTO_FAILOVER=1,
+			SOURCE_AUTO_POSITION=1,
+			SOURCE_RETRY_COUNT=%d,
+			SOURCE_CONNECT_RETRY=%d,
+			SOURCE_SSL=%d,
+			SOURCE_SSL_CA='%s',
+			SOURCE_SSL_VERIFY_SERVER_CERT=%d
+			FOR CHANNEL '%s'
+	`, replicaPass, config.Source.Host, config.Source.Port, config.SourceRetryCount, config.SourceConnectRetry, ssl, ca, sslVerify, config.Source.Name)
+
+	var errb, outb bytes.Buffer
+	err := d.exec(ctx, q, &outb, &errb)
 	if err != nil {
 		return errors.Wrapf(err, "change source for channel %s", config.Source.Name)
 	}
 
-	_, err = p.db.Exec(`START REPLICA FOR CHANNEL ?`, config.Source.Name)
+	outb.Reset()
+	errb.Reset()
+	err = d.exec(ctx, fmt.Sprintf(`START REPLICA FOR CHANNEL '%s'`, config.Source.Name), &outb, &errb)
 	return errors.Wrapf(err, "start replica for source %s", config.Source.Name)
 
 }
 
-func (p *Database) DeleteReplicationSource(name, host string, port int) error {
-	_, err := p.db.Exec("SELECT asynchronous_connection_failover_delete_source(?, ?, ?, null)", name, host, port)
+func (d *Database) DeleteReplicationSource(ctx context.Context, name, host string, port int) error {
+	var errb, outb bytes.Buffer
+	q := fmt.Sprintf("SELECT asynchronous_connection_failover_delete_source('%s', '%s', %d, null)", name, host, port)
+	err := d.exec(ctx, q, &outb, &errb)
 	return errors.Wrap(err, "delete replication source "+name)
 }
 
-func (p *Database) ProxySQLInstanceStatus(host string) ([]string, error) {
-	rows, err := p.db.Query("SELECT status FROM mysql_servers WHERE hostname LIKE ?;", host+"%")
+func (d *Database) ProxySQLInstanceStatus(ctx context.Context, host string) ([]string, error) {
+
+	rows := []*struct {
+		Status string `csv:"status"`
+	}{}
+
+	q := fmt.Sprintf("SELECT status FROM proxysql_servers WHERE hostname LIKE '%s%%'", host)
+
+	err := d.query(ctx, q, &rows)
 	if err != nil {
 		if errors.Is(err, sql.ErrNoRows) {
 			return nil, ErrNotFound
@@ -301,42 +322,46 @@ func (p *Database) ProxySQLInstanceStatus(host string) ([]string, error) {
 	}
 
 	statuses := []string{}
-	for rows.Next() {
-		var status string
-
-		err := rows.Scan(&status)
-		if err != nil {
-			return nil, err
-		}
-
-		statuses = append(statuses, status)
+	for _, row := range rows {
+		statuses = append(statuses, row.Status)
 	}
 
 	return statuses, nil
 }
 
-func (p *Database) PresentInHostgroups(host string) (bool, error) {
+func (d *Database) PresentInHostgroups(ctx context.Context, host string) (bool, error) {
 	hostgroups := []string{WriterHostgroup, ReaderHostgroup}
-	query := fmt.Sprintf(`SELECT COUNT(*) FROM mysql_servers
+
+	rows := []*struct {
+		Count int `csv:"count"`
+	}{}
+
+	q := fmt.Sprintf(`
+		SELECT COUNT(*) FROM mysql_servers
 		INNER JOIN mysql_galera_hostgroups ON hostgroup_id IN (%s)
-		WHERE hostname LIKE ? GROUP BY hostname`, strings.Join(hostgroups, ","))
-	var count int
-	err := p.db.QueryRow(query, host+"%").Scan(&count)
+		WHERE hostname LIKE '%s' GROUP BY hostname`, strings.Join(hostgroups, ","), host+"%")
+
+	err := d.query(ctx, q, &rows)
 	if err != nil {
 		if errors.Is(err, sql.ErrNoRows) {
 			return false, ErrNotFound
 		}
 		return false, err
 	}
-	if count != len(hostgroups) {
+	if rows[0].Count != len(hostgroups) {
 		return false, nil
 	}
 	return true, nil
 }
 
-func (p *Database) PrimaryHost() (string, error) {
-	var host string
-	err := p.db.QueryRow("SELECT hostname FROM runtime_mysql_servers WHERE hostgroup_id = ?", writerID).Scan(&host)
+func (d *Database) PrimaryHost(ctx context.Context) (string, error) {
+	rows := []*struct {
+		Hostname string `csv:"host"`
+	}{}
+
+	q := fmt.Sprintf("SELECT hostname as host FROM runtime_mysql_servers WHERE hostgroup_id = %d", writerID)
+
+	err := d.query(ctx, q, &rows)
 	if err != nil {
 		if errors.Is(err, sql.ErrNoRows) {
 			return "", ErrNotFound
@@ -344,12 +369,15 @@ func (p *Database) PrimaryHost() (string, error) {
 		return "", err
 	}
 
-	return host, nil
+	return rows[0].Hostname, nil
 }
 
-func (p *Database) Hostname() (string, error) {
-	var hostname string
-	err := p.db.QueryRow("SELECT @@hostname hostname").Scan(&hostname)
+func (d *Database) Hostname(ctx context.Context) (string, error) {
+	rows := []*struct {
+		Hostname string `csv:"hostname"`
+	}{}
+
+	err := d.query(ctx, "SELECT @@hostname hostname", &rows)
 	if err != nil {
 		if errors.Is(err, sql.ErrNoRows) {
 			return "", ErrNotFound
@@ -357,14 +385,16 @@ func (p *Database) Hostname() (string, error) {
 		return "", err
 	}
 
-	return hostname, nil
+	return rows[0].Hostname, nil
 }
 
-func (p *Database) WsrepLocalStateComment() (string, error) {
-	var variable_name string
-	var value string
+func (d *Database) WsrepLocalStateComment(ctx context.Context) (string, error) {
+	rows := []*struct {
+		VariableName string `csv:"Variable_name"`
+		Value        string `csv:"Value"`
+	}{}
 
-	err := p.db.QueryRow("SHOW GLOBAL STATUS LIKE 'wsrep_local_state_comment'").Scan(&variable_name, &value)
+	err := d.query(ctx, "SHOW GLOBAL STATUS LIKE 'wsrep_local_state_comment'", &rows)
 	if err != nil {
 		if err == sql.ErrNoRows {
 			return "", fmt.Errorf("variable was not found")
@@ -372,13 +402,15 @@ func (p *Database) WsrepLocalStateComment() (string, error) {
 		return "", err
 	}
 
-	return value, nil
+	return rows[0].Value, nil
 }
 
-func (p *Database) Version() (string, error) {
-	var version string
+func (d *Database) Version(ctx context.Context) (string, error) {
+	rows := []*struct {
+		Version string `csv:"version"`
+	}{}
 
-	err := p.db.QueryRow("select @@VERSION;").Scan(&version)
+	err := d.query(ctx, "select @@VERSION as version;", &rows)
 	if err != nil {
 		if err == sql.ErrNoRows {
 			return "", fmt.Errorf("variable was not found")
@@ -386,9 +418,5 @@ func (p *Database) Version() (string, error) {
 		return "", err
 	}
 
-	return version, nil
-}
-
-func (p *Database) Close() error {
-	return p.db.Close()
+	return rows[0].Version, nil
 }
